category control

+++ src/app/category/category.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { Category } from '../category';
 
 import { CategoryService } from '../category.service';
+import { Inventory } from '../inventory';
+import { InventoryService } from '../inventory.service';
 
 @Component({
   selector: 'app-category',
@@ -9,10 +11,14 @@ import { CategoryService } from '../category.service';
   styleUrls: ['./category.component.css']
 })
 export class CategoryComponent {
-  constructor(private categoryService: CategoryService) { }
+  constructor(
+    private categoryService: CategoryService,
+    private inventoryService: InventoryService) { }
 
   category: Category[] = [];
   newCategory = new Category('');
+  inventory: Inventory[] = [];
+  cleanCategory = false;
 
   ngOnInit(): void {
     this.getCategory();
@@ -30,12 +36,47 @@ export class CategoryComponent {
   }
 
   deleteCategory(cate: Category) {
-    if (confirm(`Are you sure to delete Category ${cate.category}?`)) {
-      this.categoryService.deleteCategory(cate);
-      this.getCategory();
-      window.location.reload();
-      alert('Category deleted');
-    }
+    this.checkInventories(cate.category);
+    setTimeout(() => {
+      console.log(this.cleanCategory);
+      if (confirm(`Are you sure to delete Category ${cate.category}?`)) {
+        if (this.cleanCategory === true) {
+          this.categoryService.deleteCategory(cate);
+          this.getCategory();
+          window.location.reload();
+          alert('Category deleted');
+        } else {
+          alert('You cannot delete! because there are some inventories in this category!');
+        }
+      }
+    }, 300)
+
+
   }
+
+  checkInventories(cate: any) {
+    this.inventoryService.fetchInventory().subscribe(res => {
+      this.inventory = res;
+    });
+    setTimeout(() => {
+      let found = false;
+      console.log(cate);
+      for (let key in this.inventory) {
+        if (cate === this.inventory[key].category) {
+          found = true;
+        }
+      }
+      if (found === true) {
+        console.log('found!');
+        this.cleanCategory = false;
+      } else {
+        console.log('Not found!');
+        this.cleanCategory = true;
+      }
+    }, 300);
+  }
+
+
+
 
 }
